% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_var_density.R
\name{generate_var_density}
\alias{generate_var_density}
\title{generate_var_density}
\usage{
generate_var_density(n_sample, mu_mat, std_mat, rho, seed = 1234)
}
\arguments{
\item{n_sample}{: number of samples}

\item{mu_mat}{: matrix of Gaussian means, components along rows and marginals along columns}

\item{std_mat}{: matrix of Gaussian sds, components along rows and marginals along columns}

\item{rho}{: equicorrealtion coefficient, lying betwwen -1 and 1}

\item{seed}{: seed}
}
\value{
samples from the Gaussian copula model with equicorrelation matrix
}
\description{
\code{generate_var_density} Generates samples from Gaussian Copula model with marginals specified by finite Gaussian mixtures and having equicorrelation coefficient rho
}
\examples{
generate_var_density(n_sample = 100, mu_mat = matrix(1:8, nrow = 2, byrow = TRUE),
std_mat = matrix(1, ncol = 4, nrow = 2), rho = .5, seed = 123)
}
