% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logVarDensity.R
\name{log_var_density}
\alias{log_var_density}
\title{log density and derivatives}
\usage{
log_var_density(z, mu_mat, std_mat, rho)
}
\arguments{
\item{z}{: input}

\item{mu_mat}{: matrix of Gaussian means, components along rows}

\item{std_mat}{: matrix of Gaussian sds, components along rows}

\item{rho}{: equicorrealtion coefficient, lying betwwen -1 and 1}
}
\value{
: A list containing the log density and derivatives wrt mu, std and rho
}
\description{
\code{log_var_density} Calculates the log value of the Gaussian Copula model density at the provided input, using mixture Gaussian marginals and equicorrelation coefficient rho. Also outputs derivatives of the log density wrt to all the mean and sd parameters, as well as rho
}
